restructure web in sub modules - WIP
