+++ web/all/main.go
@@ -1,0 +1,10 @@
+package all
+
+import (
+	_ "github.com/Kukoon/media-server/web/api/channel"
+	_ "github.com/Kukoon/media-server/web/api/recording"
+	_ "github.com/Kukoon/media-server/web/api/stream"
+	_ "github.com/Kukoon/media-server/web/docs"
+	_ "github.com/Kukoon/media-server/web/metrics"
+	_ "github.com/Kukoon/media-server/web/podcast"
+)
