+++ web/api/channel/main_test.go
@@ -1,4 +1,4 @@
-package web
+package channel
 
 import (
 	"net/http"
@@ -7,11 +7,13 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/Kukoon/media-server/models"
+	"github.com/Kukoon/media-server/web"
+	"github.com/Kukoon/media-server/web/webtest"
 )
 
 func TestAPIChannelList(t *testing.T) {
 	assert := assert.New(t)
-	s := testSetup(assert)
+	s := webtest.New(assert)
 	assert.NotNil(s)
 
 	list := []*models.Channel{}
@@ -21,7 +23,7 @@ func TestAPIChannelList(t *testing.T) {
 
 func TestAPIChannelGet(t *testing.T) {
 	assert := assert.New(t)
-	s := testSetup(assert)
+	s := webtest.New(assert)
 	assert.NotNil(s)
 
 	obj := models.Channel{}
@@ -32,7 +34,7 @@ func TestAPIChannelGet(t *testing.T) {
 	// GET - id
 	s.Request(http.MethodGet, "/api/v1/channel/df1555f5-7046-4f7a-adcc-195b73949723", nil, http.StatusOK, &obj)
 
-	hErr := HTTPError{}
+	hErr := web.HTTPError{}
 	// GET - not found
 	s.Request(http.MethodGet, "/api/v1/channel/00000000-0000-0000-0000-000000000001", nil, http.StatusNotFound, &hErr)
 }
