+++ web/docs/main_test.go
@@ -1,0 +1,21 @@
+package docs
+
+/* -
+import (
+	"net/http"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+
+	"github.com/Kukoon/media-server/web/webtest"
+)
+
+func TestDocsLoaded(t *testing.T) {
+	assert := assert.New(t)
+	s := webtest.New(assert)
+	assert.NotNil(s)
+
+	// GET
+	s.Request(http.MethodGet, "/api/help/index.html", nil, http.StatusOK, nil)
+}
+*/
